+++ uxn_common/src/main/java/com/example/uxn_common/global/domain/staff/repository/StaffActivityRepository.java
@@ -1,11 +1,8 @@
 package com.example.uxn_common.global.domain.staff.repository;
 
 import com.example.uxn_common.global.domain.staff.StaffActivity;
-import com.example.uxn_common.global.domain.staff.StaffAuthority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaffActivityRepository extends JpaRepository<StaffActivity, Long> {
-
-
+public interface StaffActivityRepository extends JpaRepository<StaffActivity, Integer> {
 
 }
